requirements-full.txt
# ===================================
# DDSP-SVC-Enhanced Full Requirements
# ===================================
# Complete dependencies for development and production

# ===================================
# Core DDSP-SVC Dependencies
# ===================================

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
resampy>=0.4.2
scipy>=1.11.0
numpy==1.26.4
praat-parselmouth>=0.4.3
pyworld>=0.3.4

# Machine Learning
torch>=2.0.0
torchaudio>=2.0.0
torchcrepe>=0.0.20
torchfcpe>=0.0.4
transformers>=4.35.0
scikit_learn>=1.3.0

# Configuration & Logging
PyYAML>=6.0
gin-config>=0.5.0
gin>=0.1.0
tensorboard>=2.15.0

# Fairseq (ContentVec encoder)
fairseq>=0.12.2

# Progress & Utilities
tqdm>=4.66.0
matplotlib>=3.7.0

# ===================================
# Audio Enhancement (AudioNoise)
# ===================================

# Already covered by torch, numpy, scipy

# ===================================
# Real-time Processing (Desktop GUI)
# ===================================

sounddevice>=0.4.6
FreeSimpleGUI>=5.0.0

# ===================================
# Web API & Backend
# ===================================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
starlette>=0.27.0

# Data Validation
pydantic>=2.5.0

# HTTP & Async
httpx>=0.25.0
aiofiles>=23.2.0

# ===================================
# Development Tools (Optional)
# ===================================

# Code Quality
# black>=23.0.0
# flake8>=6.1.0
# mypy>=1.7.0
# pylint>=3.0.0

# Testing
# pytest>=7.4.0
# pytest-cov>=4.1.0
# pytest-asyncio>=0.21.0

# Pre-commit Hooks
# pre-commit>=3.5.0

# ===================================
# Notes
# ===================================

# Frontend (Vue.js) dependencies are managed separately via npm
# See web/package.json for frontend requirements

# PyTorch Installation:
# For CUDA 11.8:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# For CPU only:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
